+++ src-tauri/src/db/mod.rs
@@ -1,3 +1,4 @@
+pub mod account;
 pub mod explorer;
 pub mod item;
 pub mod mapping;